api_requirements.txt
# Web Framework & Comms
fastapi
uvicorn[standard]
celery
redis

# AWS & Configuration
boto3
pydantic-settings

flower